requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet>=3.0.0

# Computer Vision
opencv-python-headless==4.9.0.80
easyocr>=1.7.0
numpy==1.26.4
scipy>=1.11.0
imutils==0.5.4

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<4.1.0

# Validation & Config
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator>=2.0.0

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# Environment
python-dotenv==1.0.1

# PDF & Excel export
reportlab==4.2.0
openpyxl==3.1.2

# Stripe (optional - for payment)
stripe==10.0.0

# OAuth verification
google-auth==2.35.0

# Background jobs / queue
celery==5.4.0
redis==5.1.1

# Tests
pytest==8.3.5
